move create action creator to the list reduxer
