+++ src/sentry/static/sentry/app/redux/actions/sharedList.js
@@ -62,6 +62,51 @@ const acSelectPage = (resource) =>
 
 const acSelect = (resource) => makeActionCreator(`SELECT_${resource}`, 'id', 'doSelect');
 
+// Create actions
+const acCreateRequest = (resource) =>
+  makeActionCreator(`CREATE_${resource}_REQUEST`, 'entry');
+
+const acCreateSuccess = (resource) =>
+  makeActionCreator(`CREATE_${resource}_SUCCESS`, 'entry');
+
+const acCreateFailure = (resource) =>
+  makeActionCreator(`CREATE_${resource}_FAILURE`, 'statusCode', 'message');
+
+const acCreate = (resource, urlTemplate) => {
+  return (org, data, onSuccess) => (dispatch) => {
+    dispatch(acCreateRequest(resource)(data));
+
+    const url = urlTemplate.replace('{org}', org.slug);
+    // use client that sentry is using
+    const api = new Client(); // TODO: use axios (must send same headers as Client does).
+    api.request(url, {
+      method: 'POST',
+      data,
+      success: (response) => {
+        dispatch(acCreateSuccess(resource)(data));
+
+        // TODO: This is to handle redirects. There is perhaps a better
+        // pattern for this using redux only?
+        if (onSuccess) {
+          onSuccess(response);
+        }
+      },
+      error: (err) => {
+        const message = getErrorMessage(err);
+        dispatch(acCreateFailure(resource)(err.status, message));
+      },
+    });
+    // TODO: uncomment this when clims-465 is completed
+    // return axios
+    //   .post(url, data)
+    //   .then(() => dispatch(acCreateSuccess(resource)(data)))
+    //   .catch((response) => {
+    //     const message = getErrorMessage(response.request);
+    //     dispatch(acCreateFailure(resource)(response.request.status, message));
+    //   });
+  };
+};
+
 function getErrorMessage(err) {
   return `(${err.statusText}) ${err.responseText}`;
 }
@@ -73,6 +118,10 @@ export const listActionCreators = {
   acGetListFailure,
   acSelect,
   acSelectPage,
+  acCreate,
+  acCreateRequest,
+  acCreateSuccess,
+  acCreateFailure,
 };
 
 // Creates all actions required for a regular resource
@@ -87,5 +136,11 @@ export const makeResourceActions = (resourceName, listUrl, getListKey = 'org') =
     // Select entries in the list or pages of it
     select: acSelect(resourceName),
     selectPage: acSelectPage(resourceName),
+
+    // Create entries
+    create: acCreate(resourceName, listUrl),
+    createRequest: acCreateRequest(resourceName),
+    createSuccess: acCreateSuccess(resourceName),
+    createFailure: acCreateFailure(resourceName),
   };
 };
