+++ src/sentry/static/sentry/app/redux/reducers/sharedEntry.js
@@ -46,29 +46,6 @@ export function getEntryFailure(state, action) {
   };
 }
 
-export function createEntryRequest(state, action) {
-  return {
-    ...state,
-    creating: true,
-  };
-}
-
-export function createEntrySuccess(state, action) {
-  return {
-    ...state,
-    creating: false,
-    entry: action.entry,
-  };
-}
-
-export function createEntryFailure(state, action) {
-  return {
-    ...state,
-    creating: false,
-    errorMessage: action.message,
-  };
-}
-
 const createResourceReducer = (resource, initialState) => (
   state = initialState,
   action
@@ -86,12 +63,6 @@ const createResourceReducer = (resource, initialState) => (
       return updateEntrySuccess(state, action);
     case `UPDATE_${resource}_FAILURE`:
       return updateEntryFailure(state, action);
-    case `CREATE_${resource}_REQUEST`:
-      return createEntryRequest(state, action);
-    case `CREATE_${resource}_SUCCESS`:
-      return createEntrySuccess(state, action);
-    case `CREATE_${resource}_FAILURE`:
-      return createEntryFailure(state, action);
     default:
       return state;
   }
@@ -103,7 +74,6 @@ export const entry = {
     loadingDetails: false,
     updating: false,
     entry: null,
-    creating: false,
   },
 
   getEntryRequest,
@@ -112,9 +82,6 @@ export const entry = {
   updateEntryRequest,
   updateEntrySuccess,
   updateEntryFailure,
-  createEntryRequest,
-  createEntrySuccess,
-  createEntryFailure,
 };
 
 export const resource = {
